requirements.txt
pyTelegramBotAPI
python-dotenv
openai
pyyaml
tiktoken